standardize methods to return json not response; jwt backwards compat fix

+++ astronomer/starship/services/remote_airflow_client.py
@@ -1,6 +1,7 @@
 import json
 from datetime import datetime, timedelta
 from typing import Any, List, Dict, Optional
+import logging
 
 import requests
 from airflow.models import Connection, Pool, Variable
@@ -27,74 +28,88 @@ def conn_to_json(connection: Connection) -> dict:
 
 
 @ttl_cache(ttl=1)
-def get_connections(deployment_url: str, token: str) -> List[Any]:
+def get_connections(deployment_url: str, token: str) -> List[Dict[str, Any]]:
+    if not deployment_url:
+        return []
     r = requests.get(
         f"{deployment_url}/api/v1/connections",
         headers={"Authorization": f"Bearer {token}"},
     )
     r.raise_for_status()
-    return r.json()["connections"]
+    return r.json().get("connections", [])
 
 
 def delete_connection(
     deployment_url: str, token: str, connection: Connection
-) -> Response:
+) -> Dict[str, Any]:
+    if not deployment_url:
+        return {}
     r = requests.delete(
         f"{deployment_url}/api/v1/connections/{connection.conn_id}",
         headers={"Authorization": f"Bearer {token}"},
     )
     r.raise_for_status()
-    return r
+    return r.json()
 
 
 def do_test_connection(
     deployment_url: str, token: str, connection: Connection
-) -> Response:
+) -> Dict[str, Any]:
+    if not deployment_url:
+        return {}
     r = requests.post(
         f"{deployment_url}/api/v1/connections/test",
         headers={"Authorization": f"Bearer {token}"},
         json=conn_to_json(connection),
     )
     r.raise_for_status()
-    return r
+    return r.json()
 
 
-def create_connection(deployment_url, token, connection) -> Response:
+def create_connection(deployment_url, token, connection) -> Dict[str, Any]:
+    if not deployment_url:
+        return {}
     r = requests.post(
         f"{deployment_url}/api/v1/connections",
         headers={"Authorization": f"Bearer {token}"},
         json=conn_to_json(connection),
     )
     r.raise_for_status()
-    return r
+    return r.json()
 
 
-def delete_pool(deployment_url, token, pool) -> Response:
+def delete_pool(deployment_url, token, pool) -> Dict[str, Any]:
+    if not deployment_url:
+        return {}
     r = requests.delete(
         f"{deployment_url}/api/v1/pools/{pool.pool}",
         headers={"Authorization": f"Bearer {token}"},
     )
     r.raise_for_status()
-    return r
+    return r.json()
 
 
-def get_pools(deployment_url: str, token: str):
+def get_pools(deployment_url: str, token: str) -> List[Dict[str, Any]]:
+    if not deployment_url:
+        return []
     r = requests.get(
         f"{deployment_url}/api/v1/pools",
         headers={"Authorization": f"Bearer {token}"},
     )
     r.raise_for_status()
-    return r.json()["pools"]
+    return r.json().get("pools", [])
 
 
-def create_pool(deployment_url, token, pool: Pool) -> Response:
+def create_pool(deployment_url, token, pool: Pool) -> Dict[str, Any]:
+    if not deployment_url:
+        return {}
     r = requests.post(
         f"{deployment_url}/api/v1/pools",
         headers={"Authorization": f"Bearer {token}"},
         json={"name": pool.pool, "slots": pool.slots, "description": pool.description},
     )
     r.raise_for_status()
-    return r
+    return r.json()
 
 
 def is_pool_migrated(deployment_url: str, token: str, pool_name: str):
@@ -102,49 +117,62 @@ def is_pool_migrated(deployment_url: str, token: str, pool_name: str):
     return pool_name in (remote_pool.get("name", "") for remote_pool in remote_pools)
 
 
-def get_variables(deployment_url: str, token: str):
+def get_variables(deployment_url: str, token: str) -> List[Dict[str, Any]]:
+    if not deployment_url:
+        return []
     r = requests.get(
         f"{deployment_url}/api/v1/variables",
         headers={"Authorization": f"Bearer {token}"},
     )
     r.raise_for_status()
-    return r.json()["variables"]
+    return r.json().get("variables", [])
 
 
-def delete_variable(deployment_url: str, token: str, variable: Variable):
+def delete_variable(
+    deployment_url: str, token: str, variable: Variable
+) -> Dict[str, Any]:
+    if not deployment_url:
+        return {}
     r = requests.delete(
         f"{deployment_url}/api/v1/variables/{variable.key}",
         headers={"Authorization": f"Bearer {token}"},
     )
     r.raise_for_status()
-    return r
+    return r.json()
 
 
 def is_variable_migrated(deployment_url: str, token: str, variable: str):
     return variable in (v["key"] for v in get_variables(deployment_url, token))
 
 
-def create_variable(deployment_url, token: str, variable: Variable):
+def create_variable(deployment_url, token: str, variable: Variable) -> Dict[str, Any]:
+    if not deployment_url:
+        return {}
     r = requests.post(
         f"{deployment_url}/api/v1/variables",
         headers={"Authorization": f"Bearer {token}"},
         json={"key": variable.key, "value": variable.val},
     )
     r.raise_for_status()
-    return r
+    return r.json()
 
 
-def set_dag_is_paused(dag_id, is_paused, deployment_url, token):
+def set_dag_is_paused(dag_id, is_paused, deployment_url, token) -> Dict[str, Any]:
+    if not deployment_url:
+        return {}
+    logging.debug(f"Attempting PATCH /dags?dag_id_pattern={dag_id}")
     r = requests.patch(
-        f"{deployment_url}/api/v1/dags?dag_id_pattern={dag_id}",
+        f"{deployment_url}/api/v1/dags/{dag_id}",
         headers={"Authorization": f"Bearer {token}"},
         json={"is_paused": is_paused},
     )
     r.raise_for_status()
-    return r
+    return r.json()
 
 
-def _get_remote_dags(deployment_url: str, token: str) -> Response:
+def _get_remote_dags(deployment_url: str, token: str) -> Optional[Response]:
+    if not deployment_url:
+        return None
     r = requests.get(
         f"{deployment_url}/api/v1/dags",
         headers={"Authorization": f"Bearer {token}"},
@@ -172,7 +200,7 @@ def get_dag(
 
     if dag_fetch_time + ttl < datetime.now():
         r = _get_remote_dags(deployment_url, token)
-        if not r.ok:
+        if not r or not r.ok:
             remote_dags = {}
             dag_fetch_time = datetime.now()
             r.raise_for_status()
@@ -192,13 +220,15 @@ def get_dag(
         return remote_dags.get(dag_id)
 
 
-def get_dag_runs(dag_id, deployment_url, token) -> Response:
+def get_dag_runs(dag_id, deployment_url, token) -> Dict[str, Any]:
+    if not deployment_url:
+        return {}
     r = requests.get(
         f"{deployment_url}/api/v1/dags/{dag_id}/dagRuns?limit=1",
         headers={"Authorization": f"Bearer {token}"},
     )
     r.raise_for_status()
-    return r
+    return r.json()
 
 
 def migrate_dag(dag: str, deployment_url: str, token: str):
